# Split crate to library and binary