--- a/Cargo.toml
+++ b/Cargo.toml
@@ -12,14 +12,23 @@ keywords = [
     "lcs",
 ]
 
+[lib]
+name = "lcs_image_diff"
+path = "src/lib.rs"
+
 [[bin]]
 name = "lcs-image-diff"
 path = "src/main.rs"
+required-features = ["binary"]
+
+[features]
+default = ["binary"]
+binary = ["futures", "futures-cpupool"]
 
 [dependencies]
 base64 = "0.9.0"
 clap = "2.30.0"
-futures = "0.1.18"
-futures-cpupool = "0.1.8"
+futures = { version = "0.1.18", optional = true }
+futures-cpupool = { version = "0.1.8", optional = true }
 image = "0.18.0"
 lcs-diff = "0.1.1"