--- a/Cargo.toml
+++ b/Cargo.toml
@@ -23,12 +23,26 @@ required-features = ["binary"]
 
 [features]
 default = ["binary"]
-binary = ["futures", "futures-cpupool"]
+binary = ["futures", "futures-cpupool", "all_image_formats"]
+all_image_formats = [
+    "image/gif_codec",
+    "image/jpeg",
+    "image/ico",
+    "image/png_codec",
+    "image/pnm",
+    "image/tga",
+    "image/tiff",
+    "image/webp",
+    "image/bmp",
+    "image/hdr",
+    "image/dxt",
+    "image/jpeg_rayon"
+]
 
 [dependencies]
 base64 = "0.9.0"
 clap = "2.30.0"
 futures = { version = "0.1.18", optional = true }
 futures-cpupool = { version = "0.1.8", optional = true }
-image = "0.18.0"
+image = { version = "0.20", default-features = false }
 lcs-diff = "0.1.1"