This will allow to reuse the code in other applications/libraries.

I added some `features` in order to minimize the dependencies of the library itself and added them to `default` so `cargo install` still works as before.

Marking original images now runs in the same thread with diff but saving images to files are still in their own threads like in #2 and saving diff also has it's own thread now. Looks like it hasn't decreased performance:

**before**:
```
$ for i in {0..2}; do time ./target/release/lcs-image-diff before.png after.png diff.png; done

real	0m1,108s
user	0m1,266s
sys	0m0,099s

real	0m1,154s
user	0m1,157s
sys	0m0,117s

real	0m1,084s
user	0m1,258s
sys	0m0,104s
```
**after**:
```
$ for i in {0..2}; do time ./target/release/lcs-image-diff before.png after.png diff.png; done

real	0m1,058s
user	0m1,442s
sys	0m0,122s

real	0m0,972s
user	0m1,510s
sys	0m0,133s

real	0m0,962s
user	0m1,533s
sys	0m0,123s
```